CLUBS UPDATED AFTER SECOND MIGRATIONS

+++ clubs/views.py
@@ -4,11 +4,10 @@ from rest_framework.response import Response
 from rest_framework import status
 from rest_framework_simplejwt.authentication import JWTAuthentication
 from rest_framework.permissions import IsAuthenticated
-from clubs.clubLOGICS.clublogics import *
 from users.models import Profile
 from .serializers import *
-
-class ResgisterClubViewSet(APIView):
+from .services import *
+class CreateClubViewSet(APIView):
     
     authentication_classes  = (JWTAuthentication,)
     permission_classes      = (IsAuthenticated,)
@@ -17,26 +16,32 @@ class ResgisterClubViewSet(APIView):
     
     def post(self,request,*args,**kwargs):
         
-        loggedInUser    = request.user.id
-        userClub        = request.data['games']
-        get_profile     = Profile.objects.get(user= loggedInUser)
-        verifyUser      = clubRepo(request)
-        print(get_profile.games)
-        
-        if (userClub in get_profile.games or 'G' in get_profile.games):
-                if verifyUser == 1:
-                    user        = Accounts.objects.get(id=loggedInUser)
-                    admin       = Clubs.objects.create(owner=user)
-                    serializer  = ClubSerializer(admin,data=request.data)
-                    if serializer.is_valid():
-                        serializer.save()
-                        return Response({"Club registered":serializer.data})  
-                    else:
-                        return Response({'errors':serializer.errors})
-                elif verifyUser==0:
-                    return Response ({'You are already an owner for a club'})
-        else:
-            return Response ({"Profile  doesn't match with selected game."})
+        try:
+            loggedInUser    = request.user.id
+            userClub        = request.data['games']
+            get_profile     = Profile.objects.get(user= loggedInUser)
+            verifyUser      = clubRepo(request)
+            if (userClub in get_profile.games or 'A' in get_profile.games):
+                    if verifyUser == 1:
+                        user        = Accounts.objects.get(id=loggedInUser)
+                        admin       = Clubs.objects.create(owner=user)
+                        serializer  = ClubSerializer(admin,data=request.data)
+                        if serializer.is_valid():
+                            serializer.save()
+                            updateClubHistory(request)
+                            
+                            return Response({"Club created":serializer.data})  
+                        else:
+                            return Response({'errors':serializer.errors})
+                    elif verifyUser==0:
+                        return Response ({"Condition 1":'No more than one club can be created by a single user at a time.',
+                                          "Condition 2":'Cannot be admin for any other club.',
+                                          "Condition 3":'Cannot be member for any other club.',
+                                          "Requirement":'Meet all condition by above.'})
+            else:
+                return Response ({"Profile  doesn't match with selected game."})
+        except:
+            return Response ({"Server Error"})
         
         
 class ClubUpdateDeleteViewSet(APIView):
@@ -46,19 +51,23 @@ class ClubUpdateDeleteViewSet(APIView):
     serializer_class        = ClubUpdateSerializer
     queryset                = Clubs.objects.all()
     
-    def put(self,request,pk,*args,**kwargs):
-        
-        user  = request.user.id
-        club = Clubs.objects.get(id=pk)
-        if club.owner.id == user:
-            serializer = ClubUpdateSerializer(club,data=request.data)
-            if serializer.is_valid():
-                serializer.save()
-                return Response({"updated data": serializer.data},status=status.HTTP_200_OK)
+    def put(self,request,pk):
+        try:
+            if Clubs.objects.filter(id=pk):
+                getClub = Clubs.objects.get(id=pk)
+                if getClub.owner.id == request.user.id:
+                    serializer  = ClubUpdateSerializer(getClub,data=request.data,partial=True)
+                    if serializer.is_valid():
+                        serializer.save()
+                        return Response({"updated data": serializer.data},status=status.HTTP_200_OK)
+                    else:
+                        return Response({"errors":serializer.errors},status=status.HTTP_304_NOT_MODIFIED)
+                else:
+                    return Response ({"Owner Error":"Only Owner of the Club can edit club Stats."})
             else:
-                return Response({"errors":serializer.errors},status=status.HTTP_304_NOT_MODIFIED)
-        else:
-            return Response({"Not the Owner or an admin for the club."}) 
+                return Response("Club does not Exists.")
+        except:
+            return Response({"Server Error"})
         
         
     def delete(self,request,pk):
@@ -67,9 +76,10 @@ class ClubUpdateDeleteViewSet(APIView):
         if Clubs.objects.filter(id=pk).filter(owner=user):
             club = Clubs.objects.get(id=pk)
             club.delete()
+            ClubHistoryPerUser.objects.filter(user=request.user.id).update(owner=False)
             return Response({"Deleted"})
         else:
-            return Response({"You are no the owner for the club or the Club doesnt not exist"})  
+            return Response({"You are not the owner for the club or the Club doesnt not exist"})  
           
 class ClubInfoViewSet(APIView):
     
@@ -101,54 +111,63 @@ class ClubMembershipViewSet(APIView):
     permission_classes      = (IsAuthenticated,)
     queryset                = MembershipRequest.objects.all()
     
-    ''' ** MEMBERSHIP **
-        -> Users can cancel the request at any time .
-        -> 3 Possible Responses
-    '''
+
     def post(self,request,pk,action): 
         
         try:
-            if Clubs.objects.get(id=pk):
-                
-                if MembershipResponses.objects.filter(club=pk,blocked=request.user.id):
-                    return Response({"Status":"You are blocked"})
-                else:
-                    
-                    list_created = membershipList(request,pk) ; '''CREATE LIST FOR MEMBERS'''
-                    if action == 1:
-                        user    = request.user.id 
-                        account = Accounts.objects.get(id=user)
-                        club_requested = Clubs.objects.get(id=pk)
-                        
-                        if MembershipRequest.objects.filter(club=club_requested).filter(is_active=True).filter(sender=user):  
-                            return Response ({"Status":"Request is pending"})
-                        
-                        else:
-                            if MembershipRequest.objects.filter(club=pk).filter(is_active=False).filter(sender=user):
-                                return Response({"Status":"You are already a member of the club"})
-                            else:
-                                if list_created == False:
-                                    return Response({"Owner is assigned to all roles in club by default."})
-                                else:
-                                    instance = MembershipRequest(club=club_requested,sender=account)
-                                    instance.save()
-                                    return Response({"Status":"Requested"})
-                    
-                    elif action == 0:
-                        user            = request.user.id 
-                        account         = Accounts.objects.get(id=user)
-                        club_requested  = Clubs.objects.get(id=pk)
-
-                        if MembershipRequest.objects.filter(club=club_requested).filter(is_active=True).filter(sender=user): 
-                            instance = MembershipRequest.objects.get(club=club_requested,sender=account) 
-                            instance.delete()
-                            return Response ({"Cancelled"})
-                        else:
-                            return Response(status=status.HTTP_400_BAD_REQUEST)
-                    
+            getHistoryUser = ClubHistoryPerUser.objects.get(user=request.user.id)
+            if getHistoryUser.owner == False:
+                if getHistoryUser.total_admin() <= 2:
+                    if getHistoryUser.total_membership() >= 5:
+                      return Response({"You are member in 5 clubs,exist from another club to join new"})
                     else:
-                        return Response({"error":"URL-/<int:action> || value should be 1 (Send Request) or 0 (Cancel Request)."})
-                
+                        if Clubs.objects.get(id=pk):
+                        
+                            if MembershipResponses.objects.filter(club=pk,blocked=request.user.id):
+                                return Response({"Status":"You are blocked"})
+                            else:
+                            
+                                list_created = membershipList(request,pk) ; '''CREATE LIST FOR MEMBERS'''
+                                if action == 1:
+                                    user    = request.user.id 
+                                    account = Accounts.objects.get(id=user)
+                                    club_requested = Clubs.objects.get(id=pk)
+                                
+                                    if MembershipRequest.objects.filter(club=club_requested).filter(is_active=True).filter(sender=user):  
+                                        return Response ({"Status":"Request is pending"})
+                                
+                                    else:
+                                        if MembershipRequest.objects.filter(club=pk).filter(is_active=False).filter(sender=user):
+                                            return Response({"Status":"You are already a member of the club"})
+                                        else:
+                                            if list_created == False:
+                                                return Response({"User Error":"You are the owner for the club."})
+                                            else:
+                                                instance = MembershipRequest(club=club_requested,sender=account)
+                                                instance.save()
+                                                return Response({"Status":"Requested"})
+                            
+                                elif action == 0:
+                                    user            = request.user.id 
+                                    account         = Accounts.objects.get(id=user)
+                                    club_requested  = Clubs.objects.get(id=pk)
+
+                                    if MembershipRequest.objects.filter(club=club_requested).filter(is_active=True).filter(sender=user): 
+                                        instance = MembershipRequest.objects.get(club=club_requested,sender=account) 
+                                        instance.delete()
+                                        return Response ({"Cancelled"})
+                                    else:
+                                        return Response(status=status.HTTP_400_BAD_REQUEST)
+                            
+                                else:
+                                    return Response({"error":"URL-/<int:action> || value should be 1 (Send Request) or 0 (Cancel Request)."})
+                        else:
+                            return Response({"Club doesnt exist."})
+                else:
+                    return Response({"Message":"You are an admin for a differnet club."})
+            else:
+                return Response ({"Exits":" You own a club, an owner cannot request for membership in another clubs."})      
+             
         except:
              return Response({"ERROR WITH SELECTED CLUB":"This club is either terminated or never existed."})   
          
@@ -162,25 +181,30 @@ class ClubMembershipResponse(APIView):
     
     def post(self,request,pk,user,action):
         try:
-            if MembershipResponses.objects.get(club=pk):
+            if MembershipResponses.objects.filter(club=pk):
                 club = Clubs.objects.get(id=pk)
                 if ((club.owner.id == request.user.id)or(Clubs.objects.filter(id=pk,admins=request.user.id))) :
                     account = Accounts.objects.get(id=user)
                     if action == 1:
-                        
-                        response_model_accepted = MembershipResponses.objects.get(club=pk)
-                        response_model_accepted.waiting.remove(user)
-                        response_model_accepted.accepted.add(user)
-                        
-                        if MembershipRequest.objects.get(club=pk,sender=account,is_active=True):
-                            membership_model_accepted = MembershipRequest.objects.get(club=pk,sender=account,is_active=True)
-                            membership_model_accepted.delete()
-                            club_requested = Clubs.objects.get(id=pk)
-                            new_instance_accepted = MembershipRequest(club=club_requested,sender=account,is_active=False)
-                            new_instance_accepted.save()                               
-                        club_model_accepted = Clubs.objects.get(id=pk)
-                        club_model_accepted.members.add(user)
-                        return Response({"Status":"Accepted"})
+                        getHistory  = ClubHistoryPerUser.objects.get(user=user)
+                        if (getHistory.total_membership() < 5 and getHistory.total_admin() < 2):
+                            if MembershipRequest.objects.get(club=pk,sender=account,is_active=True):
+                                membership_model_accepted = MembershipRequest.objects.get(club=pk,sender=account,is_active=True)
+                                membership_model_accepted.delete()
+                                club_requested = Clubs.objects.get(id=pk)
+                                new_instance_accepted = MembershipRequest(club=club_requested,sender=account,is_active=False)
+                                new_instance_accepted.save()
+                            response_model_accepted = MembershipResponses.objects.get(club=pk)
+                            response_model_accepted.waiting.remove(user)
+                            response_model_accepted.accepted.add(user)
+                            club_model_accepted = Clubs.objects.get(id=pk)
+                            club_model_accepted.members.add(user)
+                            updateHistory = ClubHistoryPerUser.objects.get(user=user)
+                            updateHistory.club_member.add(club_requested)
+                            return Response({"Status":"Accepted"})
+                        else:
+                            return Response({"Condition Error":"Requested User is already a member of 5 clubs.",
+                                             "Condition Error":"Requested User is Admin for more than 1 club."})
                     
                     elif action == 0:
                         
@@ -199,13 +223,13 @@ class ClubMembershipResponse(APIView):
                         return Response({"Status":"Blocked"})
             
                     else:
-                        return Response({"Invalid Response, || URL "})
+                        return Response({"Invalid Action Request":" Action request is invalid in the url. "})
                 else:
                     return Response({"Only owner and club admins can accept and decline memebership requests."})
             else:
                 return Response({"Club does not exists."})
         except:
-            return Response({"Error": "No Request to Respond or Invalid action"})
+            return Response({"Error": "Server error"})
 
 class RemoveMemberClubViewSet(APIView):
     
@@ -225,6 +249,8 @@ class RemoveMemberClubViewSet(APIView):
                         membership_responses_model.accepted.remove(removee)
                         membership_request_model    = MembershipRequest.objects.get(club=club,sender=removee)
                         membership_request_model.delete()
+                        updateHistory   = ClubHistoryPerUser.objects.get(user=removee)
+                        updateHistory.club_member.remove(club_called)
                         return Response({"Success Message":"Removed from club"})
                     else:
                         return Response({"Not Found":"Not a member"})
@@ -262,23 +288,31 @@ class ClubAdminsViewSet(APIView):
         
         try:
             club    = Clubs.objects.get(id=pk)
-            if club.owner.id == request.user.id:
-                if Clubs.objects.filter(id=pk,members=user):
-                    ClubAdmins.objects.get_or_create(club=club)
-                    addAdmin = ClubAdmins.objects.get(club=pk)
-                    print(addAdmin.admin_count())
-                    if addAdmin.admin_count() >= 5:
-                        return Response ({"Admin Limit": "Admin limit reached to add new admin ,remove an existing one."})
-                    else:
-                        addAdmin.clubAdmins.add(user)
-                        club.admins.add(user)
-                        return Response({"Admin added"})       
-                else:
-                    return Response({"Membership":"Not an member for the club to be an admin."})
+            account = Accounts.objects.get(id =user)
+            if ClubAdmins.objects.filter(club=club,clubAdmins=account):
+                return Response({"message":"Already a admin for the club."})
             else:
-                return Response({"You are not the owner for the club  inorder to add or remove a admin."})
+                if club.owner.id == request.user.id:
+                    if Clubs.objects.filter(id=pk,members=user):
+                        ClubAdmins.objects.get_or_create(club=club)
+                        addAdmin = ClubAdmins.objects.get(club=pk)
+                        if addAdmin.admin_count() >= 5:
+                            return Response ({"Admin Limit": "Admin limit reached  ,remove an existing one."})
+                        else:
+                            getHistory = ClubHistoryPerUser.objects.get(user=user)
+                            if getHistory.total_admin() >= 2:
+                                return Response ({"Condition Error ":"User is already Admin for other clubs."})
+                            else:
+                                addAdmin.clubAdmins.add(user)
+                                club.admins.add(user)
+                                getHistory.club_admin.add(club)
+                                return Response({"Admin added"})       
+                    else:
+                        return Response({"Membership Error":"Only a member for the club can be added as admin."})
+                else:
+                    return Response({"Message":"You are not the owner for the club  inorder to add a admin."})  
         except:
-            return Response({"Error"})
+            return Response({"Error":"Server Error"})
         
     def delete(self,request,pk,user):
